+++ app/routes/interests.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter, Depends
+from fastapi_login.exceptions import InvalidCredentialsException
 
 from ..models.orm.interest import Interest as ORMInterest
 from ..models.pydantic.interest import Interest, InterestCreateIn
@@ -19,6 +20,8 @@ async def create_interest(
 
 
 @router.delete("/interests/{id}", tags=["Interests"], response_model=Interest)
-async def delete_interest(id: int):
+async def delete_interest(id: int, current_user: User = Depends(manager)):
     interest: ORMInterest = await ORMInterest.get(id)
+    if interest.user_id != current_user.id:
+        raise InvalidCredentialsException()
     return await interest.delete()
